mcp-server/requirements.txt
# MCP server dependencies
# mcp[cli] includes FastMCP and the CLI runner (mcp run / mcp dev)
mcp[cli]>=1.0.0

# Environment variable management
python-dotenv>=1.0.1
